+++ src/_packages/markitdown.index.ts
@@ -1,0 +1,255 @@
+// ⚠️ WARNING: This code has been generated so that any manual changes will be overwritten
+// `@promptbook/core`
+
+import { collectionToJson } from '../collection/collectionToJson';
+import { createCollectionFromJson } from '../collection/constructors/createCollectionFromJson';
+import { createCollectionFromPromise } from '../collection/constructors/createCollectionFromPromise';
+import { createCollectionFromUrl } from '../collection/constructors/createCollectionFromUrl';
+import { createSubcollection } from '../collection/constructors/createSubcollection';
+import {
+    ADMIN_EMAIL,
+    ADMIN_GITHUB_NAME,
+    CLAIM,
+    DEFAULT_BOOK_TITLE,
+    DEFAULT_BOOKS_DIRNAME,
+    DEFAULT_CSV_SETTINGS,
+    DEFAULT_EXECUTIONS_CACHE_DIRNAME,
+    DEFAULT_GET_PIPELINE_COLLECTION_FUNCTION_NAME,
+    DEFAULT_INTERMEDIATE_FILES_STRATEGY,
+    DEFAULT_IS_AUTO_INSTALLED,
+    DEFAULT_IS_VERBOSE,
+    DEFAULT_MAX_EXECUTION_ATTEMPTS,
+    DEFAULT_MAX_KNOWLEDGE_SOURCES_SCRAPING_DEPTH,
+    DEFAULT_MAX_KNOWLEDGE_SOURCES_SCRAPING_TOTAL,
+    DEFAULT_MAX_PARALLEL_COUNT,
+    DEFAULT_PIPELINE_COLLECTION_BASE_FILENAME,
+    DEFAULT_REMOTE_URL,
+    DEFAULT_REMOTE_URL_PATH,
+    DEFAULT_SCRAPE_CACHE_DIRNAME,
+    DEFAULT_TASK_TITLE,
+    LOGO_DARK_SRC,
+    LOGO_LIGHT_SRC,
+    MAX_FILENAME_LENGTH,
+    NAME,
+    SET_IS_VERBOSE,
+} from '../config';
+import { ORDER_OF_PIPELINE_JSON, RESERVED_PARAMETER_NAMES } from '../constants';
+import { compilePipeline } from '../conversion/compilePipeline';
+import { parsePipeline } from '../conversion/parsePipeline';
+import { pipelineJsonToString } from '../conversion/pipelineJsonToString';
+import { prettifyPipelineString } from '../conversion/prettify/prettifyPipelineString';
+import { extractParameterNamesFromTask } from '../conversion/utils/extractParameterNamesFromTask';
+import { validatePipeline } from '../conversion/validation/validatePipeline';
+import { CallbackInterfaceTools } from '../dialogs/callback/CallbackInterfaceTools';
+import type { CallbackInterfaceToolsOptions } from '../dialogs/callback/CallbackInterfaceToolsOptions';
+import { BoilerplateError } from '../errors/0-BoilerplateError';
+import { PROMPTBOOK_ERRORS } from '../errors/0-index';
+import { AbstractFormatError } from '../errors/AbstractFormatError';
+import { CollectionError } from '../errors/CollectionError';
+import { EnvironmentMismatchError } from '../errors/EnvironmentMismatchError';
+import { ExpectError } from '../errors/ExpectError';
+import { KnowledgeScrapeError } from '../errors/KnowledgeScrapeError';
+import { LimitReachedError } from '../errors/LimitReachedError';
+import { MissingToolsError } from '../errors/MissingToolsError';
+import { NotFoundError } from '../errors/NotFoundError';
+import { NotYetImplementedError } from '../errors/NotYetImplementedError';
+import { ParseError } from '../errors/ParseError';
+import { PipelineExecutionError } from '../errors/PipelineExecutionError';
+import { PipelineLogicError } from '../errors/PipelineLogicError';
+import { PipelineUrlError } from '../errors/PipelineUrlError';
+import { UnexpectedError } from '../errors/UnexpectedError';
+import { assertsExecutionSuccessful } from '../execution/assertsExecutionSuccessful';
+import { createPipelineExecutor } from '../execution/createPipelineExecutor/00-createPipelineExecutor';
+import { embeddingVectorToString } from '../execution/embeddingVectorToString';
+import { executionReportJsonToString } from '../execution/execution-report/executionReportJsonToString';
+import type { ExecutionReportStringOptions } from '../execution/execution-report/ExecutionReportStringOptions';
+import { ExecutionReportStringOptionsDefaults } from '../execution/execution-report/ExecutionReportStringOptions';
+import { addUsage } from '../execution/utils/addUsage';
+import { isPassingExpectations } from '../execution/utils/checkExpectations';
+import { UNCERTAIN_USAGE, ZERO_USAGE } from '../execution/utils/usage-constants';
+import { usageToHuman } from '../execution/utils/usageToHuman';
+import { usageToWorktime } from '../execution/utils/usageToWorktime';
+import { CsvFormatDefinition } from '../formats/csv/CsvFormatDefinition';
+import { CsvFormatError } from '../formats/csv/CsvFormatError';
+import { MANDATORY_CSV_SETTINGS } from '../formats/csv/CsvSettings';
+import { TextFormatDefinition } from '../formats/text/TextFormatDefinition';
+import { BoilerplateFormfactorDefinition } from '../formfactors/_boilerplate/BoilerplateFormfactorDefinition';
+import { ChatbotFormfactorDefinition } from '../formfactors/chatbot/ChatbotFormfactorDefinition';
+import { GeneratorFormfactorDefinition } from '../formfactors/generator/GeneratorFormfactorDefinition';
+import { GenericFormfactorDefinition } from '../formfactors/generic/GenericFormfactorDefinition';
+import { ImageGeneratorFormfactorDefinition } from '../formfactors/image-generator/ImageGeneratorFormfactorDefinition';
+import { FORMFACTOR_DEFINITIONS } from '../formfactors/index';
+import { MatcherFormfactorDefinition } from '../formfactors/matcher/MatcherFormfactorDefinition';
+import { SheetsFormfactorDefinition } from '../formfactors/sheets/SheetsFormfactorDefinition';
+import { TranslatorFormfactorDefinition } from '../formfactors/translator/TranslatorFormfactorDefinition';
+import { $llmToolsMetadataRegister } from '../llm-providers/_common/register/$llmToolsMetadataRegister';
+import { $llmToolsRegister } from '../llm-providers/_common/register/$llmToolsRegister';
+import { createLlmToolsFromConfiguration } from '../llm-providers/_common/register/createLlmToolsFromConfiguration';
+import { cacheLlmTools } from '../llm-providers/_common/utils/cache/cacheLlmTools';
+import { countTotalUsage } from '../llm-providers/_common/utils/count-total-usage/countTotalUsage';
+import { limitTotalUsage } from '../llm-providers/_common/utils/count-total-usage/limitTotalUsage';
+import { _AnthropicClaudeMetadataRegistration } from '../llm-providers/anthropic-claude/register-configuration';
+import { _AzureOpenAiMetadataRegistration } from '../llm-providers/azure-openai/register-configuration';
+import { _GoogleMetadataRegistration } from '../llm-providers/google/register-configuration';
+import { joinLlmExecutionTools } from '../llm-providers/multiple/joinLlmExecutionTools';
+import { MultipleLlmExecutionTools } from '../llm-providers/multiple/MultipleLlmExecutionTools';
+import {
+    _OpenAiAssistantMetadataRegistration,
+    _OpenAiMetadataRegistration,
+} from '../llm-providers/openai/register-configuration';
+import { preparePersona } from '../personas/preparePersona';
+import { book } from '../pipeline/book-notation';
+import { isValidPipelineString } from '../pipeline/isValidPipelineString';
+import { GENERIC_PIPELINE_INTERFACE } from '../pipeline/PipelineInterface/constants';
+import { getPipelineInterface } from '../pipeline/PipelineInterface/getPipelineInterface';
+import { isPipelineImplementingInterface } from '../pipeline/PipelineInterface/isPipelineImplementingInterface';
+import { isPipelineInterfacesEqual } from '../pipeline/PipelineInterface/isPipelineInterfacesEqual';
+import { EXPECTATION_UNITS } from '../pipeline/PipelineJson/Expectations';
+import { isPipelinePrepared } from '../prepare/isPipelinePrepared';
+import { preparePipeline } from '../prepare/preparePipeline';
+import { prepareTasks } from '../prepare/prepareTasks';
+import { unpreparePipeline } from '../prepare/unpreparePipeline';
+import { prepareKnowledgePieces } from '../scrapers/_common/prepareKnowledgePieces';
+import { $scrapersMetadataRegister } from '../scrapers/_common/register/$scrapersMetadataRegister';
+import { $scrapersRegister } from '../scrapers/_common/register/$scrapersRegister';
+import { makeKnowledgeSourceHandler } from '../scrapers/_common/utils/makeKnowledgeSourceHandler';
+import { _LegacyDocumentScraperMetadataRegistration } from '../scrapers/document-legacy/register-metadata';
+import { _DocumentScraperMetadataRegistration } from '../scrapers/document/register-metadata';
+import { _MarkdownScraperMetadataRegistration } from '../scrapers/markdown/register-metadata';
+import { _PdfScraperMetadataRegistration } from '../scrapers/pdf/register-metadata';
+import { _WebsiteScraperMetadataRegistration } from '../scrapers/website/register-metadata';
+import { BlackholeStorage } from '../storage/blackhole/BlackholeStorage';
+import { MemoryStorage } from '../storage/memory/MemoryStorage';
+import { PrefixStorage } from '../storage/utils/PrefixStorage';
+import { MODEL_VARIANTS } from '../types/ModelVariant';
+import { NonTaskSectionTypes, SectionTypes } from '../types/SectionType';
+import { TaskTypes } from '../types/TaskType';
+import { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION } from '../version';
+
+// Note: Exporting version from each package
+export { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION };
+
+// Note: Entities of the `@promptbook/core`
+export {
+    $llmToolsMetadataRegister,
+    $llmToolsRegister,
+    $scrapersMetadataRegister,
+    $scrapersRegister,
+    _AnthropicClaudeMetadataRegistration,
+    _AzureOpenAiMetadataRegistration,
+    _DocumentScraperMetadataRegistration,
+    _GoogleMetadataRegistration,
+    _LegacyDocumentScraperMetadataRegistration,
+    _MarkdownScraperMetadataRegistration,
+    _OpenAiAssistantMetadataRegistration,
+    _OpenAiMetadataRegistration,
+    _PdfScraperMetadataRegistration,
+    _WebsiteScraperMetadataRegistration,
+    AbstractFormatError,
+    addUsage,
+    ADMIN_EMAIL,
+    ADMIN_GITHUB_NAME,
+    assertsExecutionSuccessful,
+    BlackholeStorage,
+    BoilerplateError,
+    BoilerplateFormfactorDefinition,
+    book,
+    cacheLlmTools,
+    CallbackInterfaceTools,
+    ChatbotFormfactorDefinition,
+    CLAIM,
+    CollectionError,
+    collectionToJson,
+    compilePipeline,
+    countTotalUsage,
+    createCollectionFromJson,
+    createCollectionFromPromise,
+    createCollectionFromUrl,
+    createLlmToolsFromConfiguration,
+    createPipelineExecutor,
+    createSubcollection,
+    CsvFormatDefinition,
+    CsvFormatError,
+    DEFAULT_BOOK_TITLE,
+    DEFAULT_BOOKS_DIRNAME,
+    DEFAULT_CSV_SETTINGS,
+    DEFAULT_EXECUTIONS_CACHE_DIRNAME,
+    DEFAULT_GET_PIPELINE_COLLECTION_FUNCTION_NAME,
+    DEFAULT_INTERMEDIATE_FILES_STRATEGY,
+    DEFAULT_IS_AUTO_INSTALLED,
+    DEFAULT_IS_VERBOSE,
+    DEFAULT_MAX_EXECUTION_ATTEMPTS,
+    DEFAULT_MAX_KNOWLEDGE_SOURCES_SCRAPING_DEPTH,
+    DEFAULT_MAX_KNOWLEDGE_SOURCES_SCRAPING_TOTAL,
+    DEFAULT_MAX_PARALLEL_COUNT,
+    DEFAULT_PIPELINE_COLLECTION_BASE_FILENAME,
+    DEFAULT_REMOTE_URL,
+    DEFAULT_REMOTE_URL_PATH,
+    DEFAULT_SCRAPE_CACHE_DIRNAME,
+    DEFAULT_TASK_TITLE,
+    embeddingVectorToString,
+    EnvironmentMismatchError,
+    executionReportJsonToString,
+    ExecutionReportStringOptionsDefaults,
+    EXPECTATION_UNITS,
+    ExpectError,
+    extractParameterNamesFromTask,
+    FORMFACTOR_DEFINITIONS,
+    GeneratorFormfactorDefinition,
+    GENERIC_PIPELINE_INTERFACE,
+    GenericFormfactorDefinition,
+    getPipelineInterface,
+    ImageGeneratorFormfactorDefinition,
+    isPassingExpectations,
+    isPipelineImplementingInterface,
+    isPipelineInterfacesEqual,
+    isPipelinePrepared,
+    isValidPipelineString,
+    joinLlmExecutionTools,
+    KnowledgeScrapeError,
+    LimitReachedError,
+    limitTotalUsage,
+    LOGO_DARK_SRC,
+    LOGO_LIGHT_SRC,
+    makeKnowledgeSourceHandler,
+    MANDATORY_CSV_SETTINGS,
+    MatcherFormfactorDefinition,
+    MAX_FILENAME_LENGTH,
+    MemoryStorage,
+    MissingToolsError,
+    MODEL_VARIANTS,
+    MultipleLlmExecutionTools,
+    NAME,
+    NonTaskSectionTypes,
+    NotFoundError,
+    NotYetImplementedError,
+    ORDER_OF_PIPELINE_JSON,
+    ParseError,
+    parsePipeline,
+    PipelineExecutionError,
+    pipelineJsonToString,
+    PipelineLogicError,
+    PipelineUrlError,
+    PrefixStorage,
+    prepareKnowledgePieces,
+    preparePersona,
+    preparePipeline,
+    prepareTasks,
+    prettifyPipelineString,
+    PROMPTBOOK_ERRORS,
+    RESERVED_PARAMETER_NAMES,
+    SectionTypes,
+    SET_IS_VERBOSE,
+    SheetsFormfactorDefinition,
+    TaskTypes,
+    TextFormatDefinition,
+    TranslatorFormfactorDefinition,
+    UNCERTAIN_USAGE,
+    UnexpectedError,
+    unpreparePipeline,
+    usageToHuman,
+    usageToWorktime,
+    validatePipeline,
+    ZERO_USAGE,
+};
+export type { CallbackInterfaceToolsOptions, ExecutionReportStringOptions };