+++ src/scrapers/markitdown/register-constructor.ts
@@ -1,0 +1,19 @@
+import type { Registration } from '../../utils/$Register';
+import { $scrapersRegister } from '../_common/register/$scrapersRegister';
+import { createMarkitdownScraper } from './createMarkitdownScraper';
+
+/**
+ * Registration of known scraper
+ *
+ * Warning: This is not useful for the end user, it is just a side effect of the mechanism that handles all available known scrapers
+ *
+ * @public exported from `@promptbook/markitdown`
+ * @public exported from `@promptbook/wizzard`
+ * @public exported from `@promptbook/cli`
+ */
+export const _MarkitdownScraperRegistration: Registration = $scrapersRegister.register(createMarkitdownScraper);
+
+/**
+ * TODO: [🎶] Naming "constructor" vs "creator" vs "factory"
+ * Note: [💞] Ignore a discrepancy between file name and entity name
+ */